executor: fix hang when the analyze worker unexpectedly exits
