+++ pkg/executor/test/analyzetest/BUILD.bazel
@@ -9,7 +9,7 @@ go_test(
         "main_test.go",
     ],
     flaky = True,
-    shard_count = 49,
+    shard_count = 48,
     deps = [
         "//pkg/config",
         "//pkg/domain",
